API/requirements.txt
Flask
ultralytic
numpy
opencv-python
scikit-learn
scikit-image
flask-cors